requirements.txt
python-telegram-bot == 12.2.0
mysql-connector-python == 8.0.19
telegram == 0.0.1